fix(cr/comments): fix all cr comments
