+++ src/dashboard/apigateway/apigateway/common/contexts/__init__.py
@@ -15,9 +15,7 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 #
-from .context import (  # noqa
-    GatewayFeatureFlagContext,
-    ResourceAuthContext,
-    StageProxyHTTPContext,
-)
-from .gateway_auth import GatewayAuthContext, GatewayAuthConfig  # noqa
+from .gateway_auth import GatewayAuthConfig, GatewayAuthContext  # noqa
+from .gateway_feature_flag import GatewayFeatureFlagContext  # noqa
+from .resource_auth import ResourceAuthContext  # noqa
+from .stage_proxy_http import StageProxyHTTPContext  # noqa
