+++ src/dashboard/apigateway/apigateway/common/signature.py
@@ -103,7 +103,6 @@ class SignatureValidator(SignatureGenerator):
         except Exception:
             raise ValueError("bk_timestamp is illegal, must be a timestamp integer")
 
-        # 有效期为 24h
         if timestamp < int(time.time()) - self.timestamp_expire_seconds:
             raise ValueError("bk_timestamp has expired")
         return timestamp
