+++ src/dashboard/apigateway/apigateway/schema/instances.py
@@ -28,7 +28,6 @@ from apigateway.utils.singleton import Singleton
 SCHEMA_NAME_CONTEXT_GATEWAY_BKAUTH = "ContextAPIBKAuth"
 SCHEMA_NAME_CONTEXT_RESOURCE_BKAUTH = "ContextResourceBKAuth"
 SCHEMA_NAME_CONTEXT_STAGE_PROXY_HTTP = "ContextStageProxyHTTP"
-SCHEMA_NAME_CONTEXT_STAGE_RATE_LIMIT = "ContextStageRateLimit"
 SCHEMA_NAME_CONTEXT_GATEWAY_FEATURE_FLAG = "ContextAPIFeatureFlag"
 SCHEMA_NAME_PROXY_HTTP = "ProxyHTTP"
 SCHEMA_NAME_PROXY_MOCK = "ProxyMock"
